Trivial refactoring: Remove redundant type casting

+++ src/main/java/org/codehaus/groovy/runtime/InvokerHelper.java
@@ -287,11 +287,11 @@ public class InvokerHelper {
         }
         if (value instanceof Short) {
             Short number = (Short) value;
-            return (short) -number;
+            return -number;
         }
         if (value instanceof Byte) {
             Byte number = (Byte) value;
-            return (byte) -number;
+            return -number;
         }
         if (value instanceof ArrayList) {
             // value is a list.
